requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
fire==0.1.3
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
Jinja2==2.10
MarkupSafe==1.0
nltk==3.3
numpy==1.14.2
pandas==0.22.0
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyenchant==2.0.0
Pygments==2.2.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.13
requests==2.18.4
rpy2==2.9.3
scikit-learn==0.19.1
scipy==1.0.1
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
traitlets==4.3.2
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
